move data fetch to component for animation to work
