+++ src/styles/main.scss
@@ -78,7 +78,6 @@
   align-items: center;
 
   button {
-    width: auto;
     height: 50px;
     border-style: solid;
     border-width: 1.3px;
@@ -95,7 +94,6 @@
       cursor: pointer;
       background-color: #c8e9a0;
       transition: 0.2s ease-in-out;
-      // border: none;
     }
   }
 
@@ -119,6 +117,7 @@
           transform: scale(1.2);
           transition: 0.1s ease-in-out;
           cursor: pointer;
+          animation: spin 1s linear infinite;
         }
       }
 
@@ -130,18 +129,19 @@
   .main-action-buttons-signedin {
     margin: 20px 0 0 0;
     display: flex;
-    justify-content: space-around;
-    width: 100%;
+    justify-content: space-evenly;
+    width: 50%;
     .main-action-buttons-first-group {
       display: flex;
-      justify-content: space-evenly;
+      justify-content: space-around;
       align-items: center;
-      width: 35%;
+      width: 70%;
     }
     .main-action-buttons-second-group {
       display: flex;
       align-items: center;
-      width: 30%;
+      justify-content: center;
+
       .main-action-refresh-group {
         font-size: 17px;
         display: flex;
@@ -150,11 +150,13 @@
 
         svg {
           font-size: 68px;
+
           &:hover {
             color: #c8e9a0;
             transform: scale(1.2);
             transition: 0.1s ease-in-out;
             cursor: pointer;
+            animation: spin 2s ease infinite;
           }
         }
 
@@ -215,5 +217,14 @@
 
   100% {
     transform: translateY(25px * -6);
+  }
+}
+
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
   }
 }
